# api/oas-doc.yaml
openapi: 3.0.0
info:
  version: "1"
  title: Payoneer Payout API
  description: Authorizes Payoneer payouts
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Service health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthCheck"
  /payout/{clientRefId}:
    parameters:
      - in: path
        name: clientRefId
        required: true
        schema:
          type: string
        description: reference id of the payout
    get:
      summary: Get Payout details/status
      x-router-controller: payoutController
      operationId: getPayoutStatus
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Errror
        '400':
          description: Invalid Request
        '404':
          description: Not Found
  /payout:
    post:
      summary: Initiates Payoneer payouts
      operationId: payout
      x-router-controller: payoutController
      responses:
        "200":
          description: Successfully initiated payout
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Internal Server Error
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
        "504":
          description: Timeout Error
      requestBody:
        required: true
        x-name: postBody
        content:
          application/json:
            schema:
                type: object
                title: Payouts
                properties:
                  amount:
                    type: number
                    example: 100
                  currency:
                    type: string
                    example: USD
                  payeeId:
                    type: string
                    example: test-payee-id
components:
  schemas:
    HealthCheck:
      type: object
      title: HealthCheck
      properties:
        status:
          type: string
          example: OK
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: string            